% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{clear_context}
\alias{clear_context}
\title{Clear the render context}
\usage{
clear_context()
}
\description{
Resets the stored DependencyManager to \code{NULL}. Called via \code{on.exit()} in
\code{statdown_render()}.
}
\keyword{internal}
