% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package .onLoad Function}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The library name.}

\item{pkgname}{The package name.}
}
\description{
This function is executed when the package is loaded.
It registers the custom S3 method for htmlwidgets.
}
