% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_dm}
\alias{get_dm}
\title{Retrieve the active DependencyManager}
\usage{
get_dm()
}
\value{
The active \code{DependencyManager}.
}
\description{
Fails fast if the context has not been initialised.
}
\keyword{internal}
