% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-render.R
\name{knit_print.htmlwidget}
\alias{knit_print.htmlwidget}
\title{Custom knit_print Method for HTMLWidgets}
\usage{
\method{knit_print}{htmlwidget}(x, options, ...)
}
\arguments{
\item{x}{An htmlwidget object.}

\item{options}{A list of knitr chunk options.}

\item{...}{Additional arguments passed to downstream functions.}
}
\value{
An object of class \code{asis_output} containing the rendered HTML.
}
\description{
Registers the widget's CSS/JS dependencies with the active
\link[depkit:DependencyManager]{depkit::DependencyManager} and returns the widget HTML as a raw block.
}
